refactor: archive 디렉토리 생성
