+++ Frontend/src/pages/archive/Signup.jsx
@@ -1,7 +1,7 @@
 import { useState } from "react";
-import Header from "../components/Header";
-import Footer from "../components/Footer";
-import LoadingSpinner from "../components/LoadingSpinner";
+import Header from "../../components/Header";
+import Footer from "../../components/Footer";
+import LoadingSpinner from "../../components/LoadingSpinner";
 
 export default function Signup() {
   const [form, setForm] = useState({
